helena runner not working when using a sampler in normal enchanted surrogated workflow. Needs more debugging
